requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.0
cryptography==41.0.7
daphne==4.0.0
defusedxml==0.7.1
distlib==0.3.8
dj-database-url==2.1.0
dj-rest-auth==5.0.2
Django==4.2.8
django-allauth==0.57.0
django-axes==6.2.0
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-crum==0.7.9
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
kombu==5.3.4
msgpack==1.0.7
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
platformdirs==4.1.0
pluggy==1.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pymemcache==4.0.0
pyOpenSSL==23.3.0
pyproject-api==1.6.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
tox==4.11.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.12
zope.interface==6.1
